+++ ExperimentalTools.Tests/Features/Constructor/AddNewConstructorWithParameterRefactoringTests.cs
@@ -19,7 +19,7 @@ namespace ExperimentalTools.Tests.Features.Constructor
         protected override CodeRefactoringProvider Provider =>
             new AddNewConstructorWithParameterRefactoring(new OptionsService());
 
-        [Theory, MemberData("HasActionTestData")]
+        [Theory, MemberData(nameof(HasActionTestData))]
         public Task HasActionTest(string test, string input, string expectedOutput) =>
             RunSingleActionTestAsync(input, expectedOutput);
 
@@ -469,7 +469,7 @@ namespace HelloWorld
                 }
             };
 
-        [Theory, MemberData("NoActionTestData")]
+        [Theory, MemberData(nameof(NoActionTestData))]
         public Task NoActionTest(string test, string input) =>
             RunNoActionTestAsync(input);